Fix Visual Studio warning treated as error.
Was an implicit narrowing from size_t (64-bit) to unsigned int (32-bit).

Reviewed-on: https://swiftshader-review.googlesource.com/13469
Tested-by: Nicolas Capens <nicolascapens@google.com>

+++ src/OpenGL/libGLESv2/IndexDataManager.cpp
@@ -261,7 +261,7 @@ int recomputePrimitiveCount(GLenum mode, GLsizei count, const std::vector<GLsize
 		}
 		return vertexPerPrimitive;
 	case GL_POINTS:
-		*primitiveCount = count - restartIndices.size();
+		*primitiveCount = static_cast<unsigned int>(count - restartIndices.size());
 		return 1;
 	default:
 		UNREACHABLE(mode);
